refactor account view
