+++ lib/controllers/account_controller/AccountController.dart
@@ -4,10 +4,10 @@ import 'package:Videotheque/services/FirestoreQueries.dart';
 import 'package:Videotheque/components/alert_dialog_component.dart';
 import 'package:Videotheque/Globals.dart';
 import 'package:Videotheque/utils/Singletons.dart';
-import 'package:Videotheque/views/account_view/components/ChangeMailComponent.dart';
-import 'package:Videotheque/views/account_view/components/ChangeNameComponent.dart';
-import 'package:Videotheque/views/account_view/components/ChangePasswordComponent.dart';
-import 'package:Videotheque/views/account_view/components/RemoveAccountComponent.dart';
+import 'package:Videotheque/views/AccountView/components/ChangeMailComponent.dart';
+import 'package:Videotheque/views/AccountView/components/ChangeNameComponent.dart';
+import 'package:Videotheque/views/AccountView/components/ChangePasswordComponent.dart';
+import 'package:Videotheque/views/AccountView/components/RemoveAccountComponent.dart';
 import 'package:Videotheque/views/person_view/person_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animated_dialog/flutter_animated_dialog.dart';
