+++ lib/views/AccountView/AccountView.dart
@@ -1,8 +1,8 @@
 import 'package:Videotheque/components/divider_component.dart';
 import 'package:Videotheque/components/skeleton_textline_component.dart';
-import 'package:Videotheque/views/account_view/components/ButtonComponent.dart';
-import 'package:Videotheque/views/account_view/components/SearchModeComponent.dart';
-import 'package:Videotheque/views/account_view/components/StatsComponent.dart';
+import 'package:Videotheque/views/AccountView/components/ButtonComponent.dart';
+import 'package:Videotheque/views/AccountView/components/SearchModeComponent.dart';
+import 'package:Videotheque/views/AccountView/components/StatsComponent.dart';
 import 'package:Videotheque/views/person_view/person_view.dart';
 import 'package:community_material_icon/community_material_icon.dart';
 import 'package:flutter/material.dart';
