+++ lib/views/AccountView/components/ChangeSearchModeComponent.dart
