+++ lib/views/AccountView/components/SearchModeComponent.dart
