+++ lib/views/AccountView/components/StatsComponent.dart
